fix: removing refcell from broadcaster
