+++ src/mvba/abba/test.rs
@@ -1,6 +1,3 @@
-use std::rc::Rc;
-use std::{cell::RefCell, collections::BTreeMap};
-
 use blsttc::{SecretKey, SecretKeySet, Signature};
 use rand::thread_rng;
 
@@ -12,16 +9,17 @@ use super::{
     },
     Abba,
 };
+
 use crate::mvba::hash::Hash32;
 use crate::mvba::tag::{Domain, Tag};
-use crate::mvba::{broadcaster::Broadcaster, bundle::Bundle, NodeId};
+use crate::mvba::{broadcaster::Broadcaster, NodeId};
 
 struct TestNet {
     sec_key_set: SecretKeySet,
     abba: Abba,
     proposal_digest: Hash32,
     proposal_sig: Signature,
-    broadcaster: Rc<RefCell<Broadcaster>>,
+    broadcaster: Broadcaster,
 }
 
 impl TestNet {
@@ -43,14 +41,8 @@ impl TestNet {
         let sign_bytes = crate::mvba::vcbc::c_ready_bytes_to_sign(&tag, &proposal_digest).unwrap();
         let proposal_sig = sec_key_set.secret_key().sign(sign_bytes);
 
-        let broadcaster = Rc::new(RefCell::new(Broadcaster::new(i)));
-        let abba = Abba::new(
-            tag,
-            i,
-            sec_key_set.public_keys(),
-            sec_key_share,
-            broadcaster.clone(),
-        );
+        let broadcaster = Broadcaster::new(i);
+        let abba = Abba::new(tag, i, sec_key_set.public_keys(), sec_key_share);
 
         Self {
             sec_key_set,
@@ -103,7 +95,6 @@ impl TestNet {
 
     pub fn is_broadcasted(&self, msg: &Message) -> bool {
         self.broadcaster
-            .borrow()
             .has_gossip_message(&bincode::serialize(msg).unwrap())
     }
 }
@@ -115,7 +106,7 @@ fn test_round_votes() {
     let mut t = TestNet::new(i, j);
 
     t.abba
-        .pre_vote_one(t.proposal_digest, t.proposal_sig)
+        .pre_vote_one(t.proposal_digest, t.proposal_sig, &mut t.broadcaster)
         .unwrap();
 
     assert!(t.abba.get_pre_votes_by_round(1).unwrap().len() == 1);
@@ -131,7 +122,11 @@ fn test_should_publish_pre_vote_message() {
     let mut t = TestNet::new(i, j);
 
     t.abba
-        .pre_vote_one(t.proposal_digest, t.proposal_sig.clone())
+        .pre_vote_one(
+            t.proposal_digest,
+            t.proposal_sig.clone(),
+            &mut t.broadcaster,
+        )
         .unwrap();
 
     let just = PreVoteJustification::WithValidity(t.proposal_digest, t.proposal_sig.clone());
@@ -146,7 +141,11 @@ fn test_should_publish_main_vote_message() {
     let mut t = TestNet::new(i, j);
 
     t.abba
-        .pre_vote_one(t.proposal_digest, t.proposal_sig.clone())
+        .pre_vote_one(
+            t.proposal_digest,
+            t.proposal_sig.clone(),
+            &mut t.broadcaster,
+        )
         .unwrap();
     let just = PreVoteJustification::WithValidity(t.proposal_digest, t.proposal_sig.clone());
 
@@ -154,10 +153,10 @@ fn test_should_publish_main_vote_message() {
     let pre_vote_s = t.make_pre_vote_msg(1, Value::One, &just, &TestNet::PARTY_S);
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, pre_vote_y)
+        .receive_message(TestNet::PARTY_Y, pre_vote_y, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_S, pre_vote_s)
+        .receive_message(TestNet::PARTY_S, pre_vote_s, &mut t.broadcaster)
         .unwrap();
 
     let sign_bytes = t.abba.pre_vote_bytes_to_sign(1, &Value::One).unwrap();
@@ -179,7 +178,9 @@ fn test_ignore_messages_with_wrong_domain() {
     let mut pre_vote_x = t.make_pre_vote_msg(1, Value::One, &just, &TestNet::PARTY_B);
     pre_vote_x.tag.domain = Domain::new("another-domain", 0);
 
-    let result = t.abba.receive_message(TestNet::PARTY_B, pre_vote_x);
+    let result = t
+        .abba
+        .receive_message(TestNet::PARTY_B, pre_vote_x, &mut t.broadcaster);
     match result {
         Err(Error::InvalidMessage(msg)) => assert_eq!(
             msg,
@@ -199,7 +200,9 @@ fn test_ignore_messages_with_wrong_proposer() {
     let mut pre_vote_x = t.make_pre_vote_msg(1, Value::One, &just, &TestNet::PARTY_B);
     pre_vote_x.tag.proposer = TestNet::PARTY_B;
 
-    let result = t.abba.receive_message(TestNet::PARTY_B, pre_vote_x);
+    let result = t
+        .abba
+        .receive_message(TestNet::PARTY_B, pre_vote_x, &mut t.broadcaster);
     match result {
         Err(Error::InvalidMessage(msg)) => assert_eq!(
             msg,
@@ -225,7 +228,9 @@ fn test_absent_main_vote_round_one_invalid_justification() {
 
     let main_vote_b = t.make_main_vote_msg(1, MainVoteValue::Abstain, &just, &TestNet::PARTY_B);
 
-    let result = t.abba.receive_message(TestNet::PARTY_B, main_vote_b);
+    let result = t
+        .abba
+        .receive_message(TestNet::PARTY_B, main_vote_b, &mut t.broadcaster);
     assert!(matches!(result, Err(Error::InvalidMessage(msg))
         if msg == "invalid signature for the VCBC proposal"));
 }
@@ -251,7 +256,9 @@ fn test_pre_vote_invalid_sig_share() {
         }),
     };
 
-    let result = t.abba.receive_message(TestNet::PARTY_B, msg);
+    let result = t
+        .abba
+        .receive_message(TestNet::PARTY_B, msg, &mut t.broadcaster);
     assert!(matches!(result, Err(Error::InvalidMessage(msg))
         if msg == "invalid signature share"));
 }
@@ -269,15 +276,17 @@ fn test_double_pre_vote() {
     let pre_vote_2 = t.make_pre_vote_msg(1, Value::One, &just_1, &TestNet::PARTY_B);
 
     t.abba
-        .receive_message(TestNet::PARTY_B, pre_vote_1.clone())
+        .receive_message(TestNet::PARTY_B, pre_vote_1.clone(), &mut t.broadcaster)
         .unwrap();
 
     // Repeating the message, should not return any error
     t.abba
-        .receive_message(TestNet::PARTY_B, pre_vote_1)
+        .receive_message(TestNet::PARTY_B, pre_vote_1, &mut t.broadcaster)
         .unwrap();
 
-    let result = t.abba.receive_message(TestNet::PARTY_B, pre_vote_2);
+    let result = t
+        .abba
+        .receive_message(TestNet::PARTY_B, pre_vote_2, &mut t.broadcaster);
     assert!(matches!(result, Err(Error::InvalidMessage(msg))
         if msg == format!(
             "double pre-vote detected from {:?}", &TestNet::PARTY_B)));
@@ -296,7 +305,9 @@ fn test_pre_vote_round_1_invalid_c_final_signature() {
     let just = PreVoteJustification::WithValidity(t.proposal_digest, invalid_sig);
     let msg = t.make_pre_vote_msg(1, Value::One, &just, &TestNet::PARTY_B);
 
-    let result = t.abba.receive_message(TestNet::PARTY_B, msg);
+    let result = t
+        .abba
+        .receive_message(TestNet::PARTY_B, msg, &mut t.broadcaster);
     assert!(matches!(result, Err(Error::InvalidMessage(msg))
     if msg == *"invalid signature for the VCBC proposal"));
 }
@@ -310,7 +321,9 @@ fn test_pre_vote_round_1_invalid_value_one() {
     let just = PreVoteJustification::WithValidity(t.proposal_digest, t.proposal_sig.clone());
     let msg = t.make_pre_vote_msg(1, Value::Zero, &just, &TestNet::PARTY_B);
 
-    let result = t.abba.receive_message(TestNet::PARTY_B, msg);
+    let result = t
+        .abba
+        .receive_message(TestNet::PARTY_B, msg, &mut t.broadcaster);
     assert!(matches!(result, Err(Error::InvalidMessage(msg))
     if msg == "initial value should be one"));
 }
@@ -324,7 +337,9 @@ fn test_pre_vote_round_1_invalid_value_zero() {
     let just = PreVoteJustification::FirstRoundZero;
     let msg = t.make_pre_vote_msg(1, Value::One, &just, &TestNet::PARTY_B);
 
-    let result = t.abba.receive_message(TestNet::PARTY_B, msg);
+    let result = t
+        .abba
+        .receive_message(TestNet::PARTY_B, msg, &mut t.broadcaster);
     assert!(matches!(result, Err(Error::InvalidMessage(msg))
     if msg == "initial value should be zero"));
 }
@@ -337,7 +352,7 @@ fn test_prevent_double_pre_vote() {
 
     // -- round 1
     // First we don't have the proposal, so we pre-vote for zero
-    t.abba.pre_vote_zero().unwrap();
+    t.abba.pre_vote_zero(&mut t.broadcaster).unwrap();
 
     let round_1_pre_vote_s = t.make_pre_vote_msg(
         1,
@@ -362,10 +377,10 @@ fn test_prevent_double_pre_vote() {
     );
 
     t.abba
-        .receive_message(TestNet::PARTY_X, round_1_pre_vote_x)
+        .receive_message(TestNet::PARTY_X, round_1_pre_vote_x, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_1_pre_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_1_pre_vote_y, &mut t.broadcaster)
         .unwrap();
 
     let sign_bytes = t.abba.pre_vote_bytes_to_sign(1, &Value::Zero).unwrap();
@@ -403,11 +418,11 @@ fn test_prevent_double_pre_vote() {
     );
 
     t.abba
-        .receive_message(TestNet::PARTY_X, round_1_main_vote_x)
+        .receive_message(TestNet::PARTY_X, round_1_main_vote_x, &mut t.broadcaster)
         .unwrap();
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_1_main_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_1_main_vote_y, &mut t.broadcaster)
         .unwrap();
 
     // -- round 2
@@ -427,7 +442,7 @@ fn test_prevent_double_pre_vote() {
         t.make_pre_vote_msg(2, Value::One, &weak_validity_just, &TestNet::PARTY_X);
 
     t.abba
-        .receive_message(TestNet::PARTY_X, round_2_pre_vote_x)
+        .receive_message(TestNet::PARTY_X, round_2_pre_vote_x, &mut t.broadcaster)
         .unwrap();
 
     // We should set the weak validity data
@@ -444,7 +459,7 @@ fn test_prevent_double_pre_vote() {
 
     assert!(t
         .abba
-        .receive_message(TestNet::PARTY_B, round_1_main_vote_b)
+        .receive_message(TestNet::PARTY_B, round_1_main_vote_b, &mut t.broadcaster)
         .is_ok());
 }
 
@@ -455,7 +470,11 @@ fn test_normal_case_one_round() {
     let mut t = TestNet::new(i, j);
 
     t.abba
-        .pre_vote_one(t.proposal_digest, t.proposal_sig.clone())
+        .pre_vote_one(
+            t.proposal_digest,
+            t.proposal_sig.clone(),
+            &mut t.broadcaster,
+        )
         .unwrap();
 
     let pre_vote_just =
@@ -464,10 +483,10 @@ fn test_normal_case_one_round() {
     let pre_vote_s = t.make_pre_vote_msg(1, Value::One, &pre_vote_just, &TestNet::PARTY_S);
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, pre_vote_y)
+        .receive_message(TestNet::PARTY_Y, pre_vote_y, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_S, pre_vote_s)
+        .receive_message(TestNet::PARTY_S, pre_vote_s, &mut t.broadcaster)
         .unwrap();
 
     let sign_bytes = t.abba.pre_vote_bytes_to_sign(1, &Value::One).unwrap();
@@ -479,10 +498,10 @@ fn test_normal_case_one_round() {
         t.make_main_vote_msg(1, MainVoteValue::one(), &main_vote_just, &TestNet::PARTY_S);
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, main_vote_y)
+        .receive_message(TestNet::PARTY_Y, main_vote_y, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_S, main_vote_s)
+        .receive_message(TestNet::PARTY_S, main_vote_s, &mut t.broadcaster)
         .unwrap();
 
     assert_eq!(t.abba.decided_value.unwrap().value, Value::One);
@@ -494,7 +513,7 @@ fn test_normal_case_zero() {
     let j = TestNet::PARTY_B;
     let mut t = TestNet::new(i, j);
 
-    t.abba.pre_vote_zero().unwrap();
+    t.abba.pre_vote_zero(&mut t.broadcaster).unwrap();
 
     let round_1_just_0 = PreVoteJustification::FirstRoundZero;
 
@@ -504,10 +523,10 @@ fn test_normal_case_zero() {
         t.make_pre_vote_msg(1, Value::Zero, &round_1_just_0, &TestNet::PARTY_S);
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_1_pre_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_1_pre_vote_y, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_S, round_1_pre_vote_s)
+        .receive_message(TestNet::PARTY_S, round_1_pre_vote_s, &mut t.broadcaster)
         .unwrap();
 
     let sign_bytes = t.abba.pre_vote_bytes_to_sign(1, &Value::Zero).unwrap();
@@ -522,10 +541,10 @@ fn test_normal_case_zero() {
     assert!(t.is_broadcasted(&round_1_main_vote_x));
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_1_main_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_1_main_vote_y, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_S, round_1_main_vote_s)
+        .receive_message(TestNet::PARTY_S, round_1_main_vote_s, &mut t.broadcaster)
         .unwrap();
 
     assert_eq!(t.abba.decided_value.unwrap().value, Value::Zero);
@@ -561,7 +580,11 @@ fn test_two_rounds() {
     let mut t = TestNet::new(i, j);
 
     t.abba
-        .pre_vote_one(t.proposal_digest, t.proposal_sig.clone())
+        .pre_vote_one(
+            t.proposal_digest,
+            t.proposal_sig.clone(),
+            &mut t.broadcaster,
+        )
         .unwrap();
 
     let round_1_just_0 = PreVoteJustification::FirstRoundZero;
@@ -574,10 +597,10 @@ fn test_two_rounds() {
         t.make_pre_vote_msg(1, Value::Zero, &round_1_just_0, &TestNet::PARTY_S);
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_1_pre_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_1_pre_vote_y, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_S, round_1_pre_vote_s)
+        .receive_message(TestNet::PARTY_S, round_1_pre_vote_s, &mut t.broadcaster)
         .unwrap();
 
     let round_1_main_vote_just = MainVoteJustification::Abstain(
@@ -605,10 +628,10 @@ fn test_two_rounds() {
     assert!(t.is_broadcasted(&round_1_main_vote_x));
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_1_main_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_1_main_vote_y, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_S, round_1_main_vote_s)
+        .receive_message(TestNet::PARTY_S, round_1_main_vote_s, &mut t.broadcaster)
         .unwrap();
 
     // Round 2
@@ -622,10 +645,10 @@ fn test_two_rounds() {
     assert!(t.is_broadcasted(&round_2_pre_vote_x));
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_2_pre_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_2_pre_vote_y, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_S, round_2_pre_vote_s)
+        .receive_message(TestNet::PARTY_S, round_2_pre_vote_s, &mut t.broadcaster)
         .unwrap();
 
     let sign_bytes = t.abba.pre_vote_bytes_to_sign(2, &Value::One).unwrap();
@@ -652,10 +675,10 @@ fn test_two_rounds() {
     assert!(t.is_broadcasted(&round_2_main_vote_x));
 
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_2_main_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_2_main_vote_y, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_S, round_2_main_vote_s)
+        .receive_message(TestNet::PARTY_S, round_2_main_vote_s, &mut t.broadcaster)
         .unwrap();
 
     assert_eq!(t.abba.decided_value.unwrap().value, Value::One);
@@ -694,7 +717,11 @@ fn test_three_rounds() {
     let mut t = TestNet::new(i, j);
 
     t.abba
-        .pre_vote_one(t.proposal_digest, t.proposal_sig.clone())
+        .pre_vote_one(
+            t.proposal_digest,
+            t.proposal_sig.clone(),
+            &mut t.broadcaster,
+        )
         .unwrap();
 
     let round_1_just_0 = PreVoteJustification::FirstRoundZero;
@@ -707,10 +734,10 @@ fn test_three_rounds() {
         t.make_pre_vote_msg(1, Value::Zero, &round_1_just_0, &TestNet::PARTY_Y);
 
     t.abba
-        .receive_message(TestNet::PARTY_X, round_1_pre_vote_x)
+        .receive_message(TestNet::PARTY_X, round_1_pre_vote_x, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_1_pre_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_1_pre_vote_y, &mut t.broadcaster)
         .unwrap();
 
     let round_1_main_vote_s = t.make_main_vote_msg(
@@ -744,11 +771,11 @@ fn test_three_rounds() {
     assert!(t.is_broadcasted(&round_1_main_vote_s));
 
     t.abba
-        .receive_message(TestNet::PARTY_X, round_1_main_vote_x)
+        .receive_message(TestNet::PARTY_X, round_1_main_vote_x, &mut t.broadcaster)
         .unwrap();
 
     t.abba
-        .receive_message(TestNet::PARTY_B, round_1_main_vote_b)
+        .receive_message(TestNet::PARTY_B, round_1_main_vote_b, &mut t.broadcaster)
         .unwrap();
 
     // Round 2
@@ -766,10 +793,10 @@ fn test_three_rounds() {
     );
 
     t.abba
-        .receive_message(TestNet::PARTY_X, round_2_pre_vote_x)
+        .receive_message(TestNet::PARTY_X, round_2_pre_vote_x, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_B, round_2_pre_vote_b)
+        .receive_message(TestNet::PARTY_B, round_2_pre_vote_b, &mut t.broadcaster)
         .unwrap();
 
     let sign_bytes = t.abba.pre_vote_bytes_to_sign(2, &Value::One).unwrap();
@@ -799,10 +826,10 @@ fn test_three_rounds() {
     );
 
     t.abba
-        .receive_message(TestNet::PARTY_X, round_2_main_vote_x)
+        .receive_message(TestNet::PARTY_X, round_2_main_vote_x, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_2_main_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_2_main_vote_y, &mut t.broadcaster)
         .unwrap();
 
     // Round 3
@@ -817,10 +844,10 @@ fn test_three_rounds() {
         t.make_pre_vote_msg(3, Value::One, &weak_validity_just, &TestNet::PARTY_Y);
 
     t.abba
-        .receive_message(TestNet::PARTY_X, round_3_pre_vote_x)
+        .receive_message(TestNet::PARTY_X, round_3_pre_vote_x, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_3_pre_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_3_pre_vote_y, &mut t.broadcaster)
         .unwrap();
 
     let sign_bytes = t.abba.pre_vote_bytes_to_sign(3, &Value::One).unwrap();
@@ -848,146 +875,11 @@ fn test_three_rounds() {
     );
 
     t.abba
-        .receive_message(TestNet::PARTY_X, round_3_main_vote_x)
+        .receive_message(TestNet::PARTY_X, round_3_main_vote_x, &mut t.broadcaster)
         .unwrap();
     t.abba
-        .receive_message(TestNet::PARTY_Y, round_3_main_vote_y)
+        .receive_message(TestNet::PARTY_Y, round_3_main_vote_y, &mut t.broadcaster)
         .unwrap();
 
     assert_eq!(t.abba.decided_value.unwrap().value, Value::One);
-}
-
-struct Net {
-    tag: Tag,
-    secret_key_set: SecretKeySet,
-    nodes: BTreeMap<NodeId, Abba>,
-    queue: BTreeMap<NodeId, Vec<Bundle>>,
-}
-
-impl Net {
-    fn new(n: usize, proposer: NodeId) -> Self {
-        // we can tolerate < n/3 faults
-        let faults = n.saturating_sub(1) / 3;
-
-        // we want to require n - f signature shares but
-        // blsttc wants a threshold value where you require t + 1 shares
-        // So we just subtract 1 from n - f.
-        let threshold = (n - faults).saturating_sub(1);
-        let secret_key_set = blsttc::SecretKeySet::random(threshold, &mut rand::thread_rng());
-        let public_key_set = secret_key_set.public_keys();
-        let tag = Tag::new(Domain::new("test-domain", 0), proposer);
-
-        let nodes = BTreeMap::from_iter((1..=n).map(|node_id| {
-            let key_share = secret_key_set.secret_key_share(node_id);
-            let broadcaster = Rc::new(RefCell::new(Broadcaster::new(node_id)));
-            let vcbc = Abba::new(
-                tag.clone(),
-                node_id,
-                public_key_set.clone(),
-                key_share,
-                broadcaster,
-            );
-            (node_id, vcbc)
-        }));
-
-        Net {
-            tag,
-            secret_key_set,
-            nodes,
-            queue: Default::default(),
-        }
-    }
-
-    fn node_mut(&mut self, id: NodeId) -> &mut Abba {
-        self.nodes.get_mut(&id).unwrap()
-    }
-
-    fn enqueue_bundles_from(&mut self, id: NodeId) {
-        let (send_bundles, bcast_bundles) = {
-            let mut broadcaster = self.node_mut(id).broadcaster.borrow_mut();
-            let (bcast_bundles, send_bundles) = broadcaster.take_bundles();
-            (send_bundles, bcast_bundles)
-        };
-
-        for (recipient, bundle) in send_bundles {
-            let bundle: Bundle =
-                bincode::deserialize(&bundle).expect("Failed to deserialize bundle");
-            self.queue.entry(recipient).or_default().push(bundle);
-        }
-
-        for bundle in bcast_bundles {
-            for recipient in self.nodes.keys() {
-                let bundle: Bundle =
-                    bincode::deserialize(&bundle).expect("Failed to deserialize bundle");
-                self.queue.entry(*recipient).or_default().push(bundle);
-            }
-        }
-    }
-
-    fn drain_queue(&mut self) {
-        while !self.queue.is_empty() {
-            for (recipient, queue) in std::mem::take(&mut self.queue) {
-                let recipient_node = self.node_mut(recipient);
-
-                for bundle in queue {
-                    let msg: Message = bincode::deserialize(&bundle.payload)
-                        .expect("Failed to deserialize message");
-
-                    println!("Handling message: {msg:?}");
-                    recipient_node
-                        .receive_message(bundle.initiator, msg)
-                        .expect("Failed to receive msg");
-                }
-
-                self.enqueue_bundles_from(recipient);
-            }
-        }
-    }
-
-    #[allow(unused)]
-    fn deliver(&mut self, recipient: NodeId, index: usize) {
-        if let Some(msgs) = self.queue.get_mut(&recipient) {
-            if msgs.is_empty() {
-                return;
-            }
-            let index = index % msgs.len();
-
-            let bundle = msgs.swap_remove(index);
-            let msg: Message =
-                bincode::deserialize(&bundle.payload).expect("Failed to deserialize message");
-
-            let recipient_node = self.node_mut(recipient);
-            recipient_node
-                .receive_message(bundle.initiator, msg)
-                .expect("Failed to receive message");
-            self.enqueue_bundles_from(recipient);
-        }
-    }
-}
-
-#[test]
-fn test_net_happy_path() {
-    let proposer = 1;
-    let mut net = Net::new(4, proposer);
-
-    let proposal_digest = Hash32::calculate("test-data".as_bytes());
-    let sign_bytes = crate::mvba::vcbc::c_ready_bytes_to_sign(&net.tag, &proposal_digest).unwrap();
-    let proposal_sig = net.secret_key_set.secret_key().sign(sign_bytes);
-
-    // All nodes pre-vote one
-
-    for id in Vec::from_iter(net.nodes.keys().copied()) {
-        net.node_mut(id)
-            .pre_vote_one(proposal_digest, proposal_sig.clone())
-            .expect("Failed to pre-vote");
-
-        net.enqueue_bundles_from(id);
-    }
-
-    net.drain_queue();
-
-    for (id, node) in net.nodes {
-        println!("Checking {id}");
-        assert_eq!(node.decided_value.unwrap().value, Value::One);
-    }
 }
